+++ site/locales/pt-BR.js
@@ -20,7 +20,8 @@ module.exports = {
         'notification': 'Notification',
         'loading': 'Loading',
         'collapse': 'Collapse',
-        'pagination': 'Paginação'
+        'pagination': 'Paginação',
+        'message-box': 'MessageBox'
     },
     misc: {
         'guide': 'Guia',