+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Todas as mudanças importantes deste projeto serão documentadas nesse arquivo.
 
 Formato de CHANGELOG baseado em [Keep a Changelog](https://keepachangelog.com/pt-BR/1.0.0/) seguindo o modelo de [versionamento semântico](https://semver.org/lang/pt-BR/).
 
+## 1.2.0 - _Unreleased_
+
+### Adicionado
+- Componente `MessageBox`,
+- Componente `Pagination`  
+
 ## 1.1.0 - 15/07/2019
 
 ### Adicionado
@@ -18,8 +24,7 @@ Formato de CHANGELOG baseado em [Keep a Changelog](https://keepachangelog.com/pt
 - Input:
     - Funcionalidade `Multiple Input`
 - Table
-    - Funcionalidade `Radio`
-- Componente `Pagination`    
+    - Funcionalidade `Radio`  
 
 ### Correções
 -Table